TryFrom impl for Config
