--- vinseers-cli/src/inputs/mod.rs
@@ -1,2 +1,0 @@
-pub mod config;
-pub mod parse;
